Conflicts:
	src/main/java/uk/ac/susx/mlcl/byblo/commands/AbstractCopyCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/AbstractExternalSortCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/AllPairsCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/CountCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/ExternalCountCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/ExternalSortEntriesCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/ExternalSortEventsCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/ExternalSortInstancesCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/FilterCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/commands/SortInstancesCommand.java
	src/main/java/uk/ac/susx/mlcl/byblo/enumerators/BiMapEnumerator.java
	src/main/java/uk/ac/susx/mlcl/byblo/enumerators/DoubleEnumeratingDelegate.java
	src/main/java/uk/ac/susx/mlcl/byblo/enumerators/EnumeratingDelegate.java
	src/main/java/uk/ac/susx/mlcl/byblo/enumerators/SingleEnumeratingDelegate.java
	src/main/java/uk/ac/susx/mlcl/byblo/measures/Measures.java
	src/main/java/uk/ac/susx/mlcl/byblo/measures/impl/LpSpaceDistance.java
	src/main/java/uk/ac/susx/mlcl/byblo/tasks/CountTask.java
	src/main/java/uk/ac/susx/mlcl/byblo/tasks/ThreadedApssTask.java
	src/main/java/uk/ac/susx/mlcl/byblo/weighings/impl/Bound.java
	src/main/java/uk/ac/susx/mlcl/lib/Predicates2.java
	src/main/java/uk/ac/susx/mlcl/lib/collect/ForwardingBiMap.java
	src/main/java/uk/ac/susx/mlcl/lib/commands/AbstractCommand.java
	src/main/java/uk/ac/susx/mlcl/lib/commands/FileDelegate.java
	src/main/java/uk/ac/susx/mlcl/lib/commands/FileMoveCommand.java
	src/main/java/uk/ac/susx/mlcl/lib/commands/FilePipeDelegate.java
	src/main/java/uk/ac/susx/mlcl/lib/events/ProgressDelegate.java
	src/main/java/uk/ac/susx/mlcl/lib/tasks/FileCopyTask.java
	src/main/java/uk/ac/susx/mlcl/lib/tasks/FileDeleteTask.java
	src/main/java/uk/ac/susx/mlcl/lib/tasks/ObjectMergeTask.java
	src/test/java/uk/ac/susx/mlcl/TestConstants.java
	src/test/java/uk/ac/susx/mlcl/byblo/FullBuildTest2.java
	src/test/java/uk/ac/susx/mlcl/byblo/JDBMTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/CountCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/ExternalCountCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/ExternalSimsKnnCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/ExternalSortEventsCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/FilterCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/IndexSimsCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/IndexTPCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/commands/IndexWTPCommandTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/io/EntryTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/io/FeatureTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/measures/MeasuresTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/tasks/AllPairsTaskTest.java
	src/test/java/uk/ac/susx/mlcl/byblo/weighings/WeightingsTest.java

+++ src/test/java/uk/ac/susx/mlcl/byblo/enumerators/JDBMStringEnumeratorTest.java
@@ -30,27 +30,24 @@
  */
 package uk.ac.susx.mlcl.byblo.enumerators;
 
+import org.apache.jdbm.DBMaker;
+import org.junit.*;
+import uk.ac.susx.mlcl.lib.ZipfianDistribution;
+
 import java.io.File;
 import java.io.IOException;
 import java.text.DecimalFormat;
 import java.text.Format;
 import java.util.Random;
-import org.apache.jdbm.DBMaker;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import static uk.ac.susx.mlcl.TestConstants.*;
-import uk.ac.susx.mlcl.lib.ZipfianDistribution;
+
+import static uk.ac.susx.mlcl.TestConstants.TEST_OUTPUT_DIR;
 
 /**
- *
  * @author Hamish I A Morgan &lt;hamish.morgan@sussex.ac.uk&gt;
  */
-public class JDBCStringEnumeratorTest {
+public class JDBMStringEnumeratorTest {
 
-    public JDBCStringEnumeratorTest() {
+    public JDBMStringEnumeratorTest() {
     }
 
     static int populationSize;
@@ -67,7 +64,7 @@ public class JDBCStringEnumeratorTest {
         zipfExponent = 1;
         repeats = 10000;
         zipfDist = new ZipfianDistribution(populationSize, zipfExponent);
-        new JDBCStringEnumeratorTest().performanceTest_JDBM_trans_mru100();
+        new JDBMStringEnumeratorTest().performanceTest_JDBM_trans_mru100();
     }
 
     @AfterClass
@@ -235,7 +232,7 @@ public class JDBCStringEnumeratorTest {
     }
 
     private void performanceTest_JDBM(
-            boolean memory, boolean trans, CacheType cacheType, int mruSize,
+            boolean memory, boolean trans, final CacheType cacheType, int mruSize,
             boolean disableCacheAutoClear, boolean disableLocking) throws IOException {
 
         final DBMaker maker;
@@ -246,7 +243,8 @@ public class JDBCStringEnumeratorTest {
             file = null;
         } else {
             file = File.createTempFile("jdbmtest", "", TEST_OUTPUT_DIR);
-            file.delete();
+            if (!file.delete())
+                throw new IOException("Unable to delete file " + file);
             maker = DBMaker.openFile(file.toString());
             maker.deleteFilesAfterClose();
 
@@ -308,7 +306,7 @@ public class JDBCStringEnumeratorTest {
         Format fmt = new DecimalFormat("%010d");
 
         for (int i = 0; i < repeats; i++) {
-            double u = zipfDist.random();
+            zipfDist.random();
             idx.indexOf(fmt.format(zipfDist.random()));
         }
 
@@ -319,10 +317,9 @@ public class JDBCStringEnumeratorTest {
 
 
         for (int i = 0; i < repeats; i++) {
-            double u = zipfDist.random();
+            zipfDist.random();
             idx.indexOf(fmt.format(zipfDist.random()));
         }
 
     }
-
 }
